+++ src/main/java/com/github/alebabai/lindenhoney/repository/VerseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.data.rest.core.annotation.RestResource;
 
 @RepositoryRestResource(path = "verses", collectionResourceRel = "verses")
-public interface VerseRepository extends PagingAndSortingRepository<Verse, Long> {
+public interface VerseRepository extends PagingAndSortingRepository<Verse, Integer> {
 
     @RestResource(exported = false)
     @Override
@@ -14,5 +14,5 @@ public interface VerseRepository extends PagingAndSortingRepository<Verse, Long>
 
     @RestResource(exported = false)
     @Override
-    void delete(Long id);
+    void delete(Integer id);
 }
